clean up poms move pubsub container management to AbstractIlpOverHttpIT remove unused event properties
