+++ connector-service-api/src/main/java/org/interledger/connector/gcp/GcpFulfillmentEvent.java
@@ -41,10 +41,6 @@ public interface GcpFulfillmentEvent {
 
   BigDecimal exchangeRate();
 
-  Optional<String> senderPaymentPointer();
-
-  Optional<String> destinationPaymentPointer();
-
   InterledgerAddress connectorIlpAddress();
 
   InterledgerAddress destinationIlpAddress();
