+++ pom.xml
@@ -26,8 +26,7 @@
     <problem.version>0.23.0</problem.version>
     <problem-spring-web.version>0.25.2</problem-spring-web.version>
     <failBuildOnAnyVulnerability>false</failBuildOnAnyVulnerability>
-    <spring-cloud.version>Greenwich.SR2</spring-cloud.version>
-    <!-- <spring-cloud.version>Greenwich.SR1</spring-cloud.version> -->
+    <spring-cloud.version>Greenwich.SR3</spring-cloud.version>
     <resilience4j.version>1.1.0</resilience4j.version>
     <moneta.version>1.3</moneta.version>
     <prometheus.version>0.8.0</prometheus.version>
@@ -186,6 +185,11 @@
         <groupId>com.google.api</groupId>
         <artifactId>gax-grpc</artifactId>
         <version>1.50.1</version> <!-- transitive via google-cloud-kms but defaults to 1.45.0 which has conflicts -->
+      </dependency>
+      <dependency>
+        <groupId>io.grpc</groupId>
+        <artifactId>grpc-api</artifactId>
+        <version>1.25.0</version> <!-- this should match the other grpc dependency versions in spring-cloud bom -->
       </dependency>
       <dependency>
         <groupId>com.google.guava</groupId>
